+++ lib/core/routes/routes.dart
@@ -1,6 +1,6 @@
 import 'package:go_router/go_router.dart';
 import 'package:sozzle/src/home/home.dart';
-import 'package:sozzle/src/settings/view/settings-page.dart';
+import 'package:sozzle/src/settings/view/settings_page.dart';
 import 'package:sozzle/src/splash/splash_page.dart';
 
 final GoRouter router = GoRouter(
